+++ src/main/scala/edu/ie3/simona/event/listener/Transformer3wResultSupport.scala
@@ -118,7 +118,7 @@ private[listener] trait Transformer3wResultSupport {
     }
   }
 
-  final object AggregatedTransformer3wResult {
+  object AggregatedTransformer3wResult {
     val EMPTY: AggregatedTransformer3wResult =
       AggregatedTransformer3wResult(None, None, None)
   }