+++ src/main/scala/edu/ie3/simona/io/runtime/RuntimeEventLogSink.scala
@@ -7,7 +7,7 @@
 package edu.ie3.simona.io.runtime
 
 import edu.ie3.simona.event.RuntimeEvent
-import edu.ie3.simona.event.RuntimeEvent._
+import edu.ie3.simona.event.RuntimeEvent.*
 import edu.ie3.simona.io.runtime.RuntimeEventSink.RuntimeStats
 import edu.ie3.simona.util.TickUtil.TickLong
 import edu.ie3.util.TimeUtil
@@ -54,7 +54,7 @@ final case class RuntimeEventLogSink(
 
       case Done(currentTick, duration, errorInSim) =>
         val simStatus =
-          if (errorInSim)
+          if errorInSim then
             s"\u001b[0;31mERROR (Failed PF: ${runtimeStats.failedPowerFlows})\u001b[0;0m"
           else
             s"\u001b[0;32mSUCCESS (Failed PF: ${runtimeStats.failedPowerFlows})\u001b[0;0m"
@@ -73,7 +73,7 @@ final case class RuntimeEventLogSink(
 
   private def calcTime(currentTick: Long): String = {
     TimeUtil.withDefaults.toString(
-      currentTick.toDateTime(
+      currentTick.toDateTime(using
         simulationStartDate
       )
     )