+++ src/main/scala/edu/ie3/simona/model/participant/FixedFeedInModel.scala
@@ -19,10 +19,9 @@ import edu.ie3.simona.service.Data.PrimaryData.{
   PrimaryDataWithComplexPower,
 }
 import edu.ie3.simona.service.ServiceType
-import edu.ie3.util.quantities.PowerSystemUnits
 import edu.ie3.util.quantities.QuantityUtils.{asMegaVar, asMegaWatt}
 import edu.ie3.util.scala.quantities.ApparentPower
-import edu.ie3.util.scala.quantities.QuantityConversionUtils.PowerConversionSimona
+import edu.ie3.util.scala.quantities.QuantityConversionUtils.toApparent
 import squants.Power
 
 import java.time.ZonedDateTime