+++ src/main/scala/edu/ie3/simona/model/participant/storage/StorageModel.scala
@@ -38,9 +38,8 @@ import edu.ie3.util.quantities.QuantityUtils.{asMegaVar, asMegaWatt, asPu}
 import edu.ie3.util.scala.quantities.ApparentPower
 import edu.ie3.util.scala.quantities.DefaultQuantities.{zeroKW, zeroKWh}
 import edu.ie3.util.scala.quantities.QuantityConversionUtils.{
-  DimensionlessToSimona,
-  EnergyToSimona,
-  PowerConversionSimona,
+  toApparent,
+  toSquants,
 }
 import squants.{Dimensionless, Energy, Power, Seconds}
 
@@ -182,17 +181,15 @@ class StorageModel private (
       setPower: Power,
   ): (ActivePowerOperatingPoint, ParticipantModel.OperationChangeIndicator) = {
     val adaptedSetPower =
-      if (
+      if
         // if power is close to zero, set it to zero
         (setPower ~= zeroKW)
         // do not keep charging if we're already full (including safety margin)
         || (setPower > zeroKW && isFull(state.storedEnergy))
         // do not keep discharging if we're already empty (including safety margin)
         || (setPower < zeroKW && isEmpty(state.storedEnergy))
-      )
-        zeroKW
-      else
-        setPower
+      then zeroKW
+      else setPower
 
     // if the storage is at minimum or maximum charged energy AND we are charging
     // or discharging, flex options will be different at the next activation