+++ src/main/scala/edu/ie3/simona/model/thermal/ThermalGrid.scala
@@ -32,7 +32,7 @@ import edu.ie3.util.quantities.QuantityUtils.{
   asPu,
   asMegaWatt,
 }
-import edu.ie3.util.scala.quantities.DefaultQuantities._
+import edu.ie3.util.scala.quantities.DefaultQuantities.*
 import squants.energy.KilowattHours
 import squants.{Energy, Power, Temperature}
 
@@ -112,7 +112,7 @@ final case class ThermalGrid(
 
     val houseDemand = house.zip(thermalGridState.houseState) match {
       case Some((thermalHouse, houseState)) =>
-        if (houseState.innerTemperature < thermalHouse.targetTemperature) {
+        if houseState.innerTemperature < thermalHouse.targetTemperature then {
           thermalHouse.energyDemand(houseState)
         } else {
           ThermalEnergyDemand.noDemand
@@ -125,10 +125,8 @@ final case class ThermalGrid(
         case Some((storage, storageState)) =>
           val storedEnergy = storageState.storedEnergy
           val storageRequired = {
-            if (storedEnergy == zeroKWh)
-              storage.getMaxEnergyThreshold
-            else
-              zeroMWh
+            if storedEnergy == zeroKWh then storage.getMaxEnergyThreshold
+            else zeroMWh
           }
 
           val storagePossible = storage.getMaxEnergyThreshold - storedEnergy
@@ -176,13 +174,10 @@ final case class ThermalGrid(
     /* Consider the action in the last state
     We can continue using the qDots from last operating point to keep continuity.
     If the house was heated in lastState and has still some demand. */
-    if (
-      state.lastHpOperatingPoint.thermalOps.qDotHouse > zeroKW && state.thermalDemands.houseDemand.hasPossibleDemand
-    )
-      handleCase(state, qDot, zeroKW)
+    if state.lastHpOperatingPoint.thermalOps.qDotHouse > zeroKW && state.thermalDemands.houseDemand.hasPossibleDemand
+    then handleCase(state, qDot, zeroKW)
     // or finally check for all other cases.
-    else
-      handleFinalFeedInCases(state, qDot)
+    else handleFinalFeedInCases(state, qDot)
   }
 
   /** Handles the last cases of [[ThermalGrid.handleFeedIn]], where the thermal
@@ -229,16 +224,13 @@ final case class ThermalGrid(
       qDot: Power,
   ): (ThermalGridOperatingPoint, Option[ThermalThreshold]) = {
 
-    if (state.thermalDemands.houseDemand.hasRequiredDemand)
+    if state.thermalDemands.houseDemand.hasRequiredDemand then
       handleCase(state, qDot, zeroKW)
-    else if (
-      state.thermalDemands.heatStorageDemand.hasRequiredDemand || state.thermalDemands.heatStorageDemand.hasPossibleDemand
-    )
-      handleCase(state, zeroKW, qDot)
-    else if (state.thermalDemands.houseDemand.hasPossibleDemand)
+    else if state.thermalDemands.heatStorageDemand.hasRequiredDemand || state.thermalDemands.heatStorageDemand.hasPossibleDemand
+    then handleCase(state, zeroKW, qDot)
+    else if state.thermalDemands.houseDemand.hasPossibleDemand then
       handleCase(state, qDot, zeroKW)
-    else
-      handleCase(state, zeroKW, zeroKW)
+    else handleCase(state, zeroKW, zeroKW)
   }
 
   /** Handles the different thermal flows from and into the thermal grid.
@@ -265,9 +257,11 @@ final case class ThermalGrid(
     val heatStorageThreshold =
       handleFeedInStorage(state, qDotHeatStorage)
 
-    val nextThreshold = determineMostRecentThreshold(
-      thermalHouseThreshold,
-      heatStorageThreshold,
+    val nextThreshold = determineNextThreshold(
+      Seq(
+        thermalHouseThreshold,
+        heatStorageThreshold,
+      )
     )
 
     (
@@ -297,11 +291,10 @@ final case class ThermalGrid(
     house.zip(state.thermalGridState.houseState) match {
       case Some((thermalHouse, houseState)) =>
         /* Check if house can handle the thermal feed in */
-        if (
-          thermalHouse.isInnerTemperatureTooHigh(
+        if thermalHouse.isInnerTemperatureTooHigh(
             houseState.innerTemperature
           )
-        ) {
+        then {
 
           val maybeFullHouseThreshold =
             thermalHouse.determineNextThreshold(houseState, zeroKW)
@@ -343,28 +336,25 @@ final case class ThermalGrid(
     }
   }
 
-  /** Determines the most recent threshold of two given input thresholds.
+  /** Determines the next threshold of a given input sequence of thresholds.
     *
-    * @param maybeHouseThreshold
-    *   Option of a possible next threshold of the thermal house.
-    * @param maybeStorageThreshold
-    *   Option of a possible next threshold of the thermal storage.
+    * @param thresholds
+    *   Sequence of Options of possible next thresholds from the thermal house
+    *   or storage.
     * @return
     *   The next threshold.
     */
-  private def determineMostRecentThreshold(
-      maybeHouseThreshold: Option[ThermalThreshold],
-      maybeStorageThreshold: Option[ThermalThreshold],
+  private def determineNextThreshold(
+      thresholds: Seq[Option[ThermalThreshold]]
   ): Option[ThermalThreshold] =
-    (maybeHouseThreshold, maybeStorageThreshold) match {
-      case (Some(houseThreshold), Some(storageThreshold)) =>
-        if (houseThreshold.tick <= storageThreshold.tick)
-          maybeHouseThreshold
-        else
-          maybeStorageThreshold
-      case (None, Some(_)) => maybeStorageThreshold
-      case (Some(_), None) => maybeHouseThreshold
-      case _               => None
+    thresholds.flatten.foldLeft[Option[ThermalThreshold]](None) {
+      case (currentMin, threshold) =>
+        currentMin match {
+          case None => Some(threshold)
+          case Some(minThreshold) =>
+            if threshold.tick < minThreshold.tick then Some(threshold)
+            else currentMin
+        }
     }
 
   /** Handle consumption (or no feed in) from thermal grid.
@@ -430,9 +420,11 @@ final case class ThermalGrid(
         storageState,
         thermalStorage.getpThermalMax * -1,
       )
-      val nextThreshold = determineMostRecentThreshold(
-        revisedHouseThreshold,
-        revisedStorageThreshold,
+      val nextThreshold = determineNextThreshold(
+        Seq(
+          revisedHouseThreshold,
+          revisedStorageThreshold,
+        )
       )
 
       (
@@ -653,14 +645,13 @@ object ThermalGrid {
         required: Energy,
         possible: Energy,
     ): ThermalEnergyDemand = {
-      if (
-        math.abs(possible.toKilowattHours) < math.abs(required.toKilowattHours)
-      )
+      if math.abs(possible.toKilowattHours) < math.abs(required.toKilowattHours)
+      then
         throw new InvalidParameterException(
           s"The possible amount of energy $possible is smaller than the required amount of energy $required. This is not supported."
         )
 
-      if (possible.toKilowattHours < 0 || required.toKilowattHours < 0)
+      if possible.toKilowattHours < 0 || required.toKilowattHours < 0 then
         throw new InvalidParameterException(
           s"The possible $possible or required $required amount of energy cannot be negative. This is not supported."
         )