+++ src/main/scala/edu/ie3/simona/service/ev/ExtEvDataService.scala
@@ -22,7 +22,7 @@ import edu.ie3.simona.exceptions.{
   ServiceException,
 }
 import edu.ie3.simona.model.participant.evcs.EvModelWrapper
-import edu.ie3.simona.ontology.messages.{Activation, ServiceMessage}
+import edu.ie3.simona.ontology.messages.ServiceMessage
 import edu.ie3.simona.ontology.messages.ServiceMessage.*
 import edu.ie3.simona.service.Data.SecondaryData.ArrivingEvs
 import edu.ie3.simona.service.ServiceStateData.{
@@ -60,7 +60,7 @@ object ExtEvDataService extends SimonaService with ExtDataSupport {
 
   override def init(
       initServiceData: ServiceStateData.InitializeServiceStateData
-  ): Try[
+  )(using log: Logger): Try[
     (
         ExtEvStateData,
         Option[Long],
@@ -162,7 +162,6 @@ object ExtEvDataService extends SimonaService with ExtDataSupport {
       map.asScala.view.mapValues(_.asScala.toSeq).toMap
 
     given context: ActorContext[Message] = ctx
-    given log: Logger = ctx.log
 
     serviceStateData.extEvMessage
       .map {
@@ -222,7 +221,7 @@ object ExtEvDataService extends SimonaService with ExtDataSupport {
       }.toSet
 
     // if there are no evcs, we're sending response right away
-    if (freeLots.isEmpty)
+    if freeLots.isEmpty then
       serviceStateData.extEvData.queueExtResponseMsg(new ProvideEvcsFreeLots())
 
     (
@@ -262,7 +261,7 @@ object ExtEvDataService extends SimonaService with ExtDataSupport {
 
     // if there are no departing evs during this tick,
     // we're sending response right away
-    if (departingEvResponses.isEmpty)
+    if departingEvResponses.isEmpty then
       serviceStateData.extEvData.queueExtResponseMsg(new ProvideDepartingEvs())
 
     (
@@ -283,7 +282,7 @@ object ExtEvDataService extends SimonaService with ExtDataSupport {
       ctx: ActorContext[Message],
   ): (ExtEvStateData, Option[Long]) = {
 
-    if (tick == INIT_SIM_TICK) {
+    if tick == INIT_SIM_TICK then {
       // During initialization, an empty ProvideArrivingEvs message
       // is sent, which includes the first relevant tick
 
@@ -341,7 +340,7 @@ object ExtEvDataService extends SimonaService with ExtDataSupport {
         val updatedResponses =
           serviceStateData.departingEvResponses.addData(evcs, evModels)
 
-        if (updatedResponses.nonComplete) {
+        if updatedResponses.nonComplete then {
           // responses are still incomplete
           serviceStateData.copy(
             departingEvResponses = updatedResponses
@@ -362,7 +361,7 @@ object ExtEvDataService extends SimonaService with ExtDataSupport {
       case FreeLotsResponse(evcs, freeLots) =>
         val updatedFreeLots = serviceStateData.freeLots.addData(evcs, freeLots)
 
-        if (updatedFreeLots.nonComplete) {
+        if updatedFreeLots.nonComplete then {
           // responses are still incomplete
           serviceStateData.copy(
             freeLots = updatedFreeLots